Better Error Messages, Support for React on Rails Pro
- Errors raised are of type ReactOnRailsError.
- Unused `server_render_method` was removed from the configuration.
- Removed ReactOnRails::Utils.server_bundle_file_name and
  ReactOnRails::Utils.bundle_file_name.
