+++ lib/react_on_rails.rb
@@ -2,6 +2,8 @@
 
 require "rails"
 
+require "react_on_rails/error"
+require "react_on_rails/react_on_rails_helper"
 require "react_on_rails/controller"
 require "react_on_rails/version"
 require "react_on_rails/version_checker"
@@ -13,9 +15,8 @@ require "react_on_rails/version_syntax_converter"
 require "react_on_rails/test_helper"
 require "react_on_rails/git_utils"
 require "react_on_rails/utils"
-require "react_on_rails/test_helper"
+require "react_on_rails/webpacker_utils"
 require "react_on_rails/test_helper/webpack_assets_compiler"
 require "react_on_rails/test_helper/webpack_assets_status_checker"
 require "react_on_rails/test_helper/ensure_assets_compiled"
-require "react_on_rails/test_helper/node_process_launcher"
 require "react_on_rails/locales_to_js"
