+++ lib/react_on_rails/error.rb
@@ -1,0 +1,4 @@
+module ReactOnRails
+  class Error < StandardError
+  end
+end
