--- lib/react_on_rails/test_helper/node_process_launcher.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module ReactOnRails
-  module TestHelper
-    def self.launch_node
-      return unless ReactOnRails.configuration.server_render_method == "NodeJS"
-
-      path = "#{::Rails.root}/client/node"
-      puts "Launching NodeJS server at #{path}"
-      system("cd #{path} && yarn start &")
-      sleep(1)
-    end
-  end
-end
