+++ spec/dummy/Gemfile
@@ -12,7 +12,7 @@ gem "webpacker", "3.0.2"
 # gem 'webpacker', path: "../../../../forks/webpacker"
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem "rails", "~> 5.1.2"
+gem "rails", "~> 5.1.6"
 # Use sqlite3 as the database for Active Record
 gem "sqlite3"
 # Use SCSS for stylesheets
