+++ spec/react_on_rails/utils_spec.rb
@@ -1,50 +1,9 @@
 # frozen_string_literal: true
 
-# rubocop:disable Metrics/ModuleLength
-# rubocop:disable Metrics/BlockLength
-
 require_relative "spec_helper"
 
 module ReactOnRails
   RSpec.describe Utils do
-    describe "cache helpers .server_bundle_file_name and .bundle_file_name" do
-      context "and file in manifest", :webpacker do
-        before do
-          allow(Rails).to receive(:root).and_return(Pathname.new("."))
-          allow(ReactOnRails).to receive_message_chain("configuration.generated_assets_dir")
-            .and_return("public/webpack/production")
-          allow(Webpacker).to receive_message_chain("config.public_output_path")
-            .and_return("public/webpack/production")
-          allow(Utils).to receive(:using_webpacker?).and_return(true)
-        end
-        describe ".bundle_file_name" do
-          before do
-            allow(Webpacker).to receive_message_chain("manifest.lookup")
-              .with("client-bundle.js")
-              .and_return("/webpack/production/client-bundle-0123456789abcdef.js")
-          end
-          subject do
-            Utils.bundle_file_name("client-bundle.js")
-          end
-          it { expect(subject).to eq("client-bundle-0123456789abcdef.js") }
-        end
-
-        describe ".server_bundle_file_name" do
-          before do
-            allow(ReactOnRails).to receive_message_chain("configuration.server_bundle_js_file")
-              .and_return("server-bundle.js")
-            allow(Webpacker).to receive_message_chain("manifest.lookup")
-              .with("server-bundle.js")
-              .and_return("/webpack/production/server-bundle-0123456789abcdef.js")
-          end
-          subject do
-            Utils.server_bundle_file_name
-          end
-          it { expect(subject).to eq("server-bundle-0123456789abcdef.js") }
-        end
-      end
-    end
-
     describe ".bundle_js_file_path" do
       before do
         allow(ReactOnRails).to receive_message_chain(:configuration, :generated_assets_dir)
@@ -62,7 +21,7 @@ module ReactOnRails
             .and_return(false)
           allow(Webpacker).to receive_message_chain("config.public_output_path")
             .and_return("/webpack/development")
-          allow(Utils).to receive(:using_webpacker?).and_return(true)
+          allow(ReactOnRails::WebpackerUtils).to receive(:using_webpacker?).and_return(true)
         end
 
         context "and file in manifest", :webpacker do
@@ -85,7 +44,7 @@ module ReactOnRails
       end
 
       context "Without Webpacker enabled" do
-        before { allow(Utils).to receive(:using_webpacker?).and_return(false) }
+        before { allow(ReactOnRails::WebpackerUtils).to receive(:using_webpacker?).and_return(false) }
 
         it { expect(subject).to eq("public/webpack/development/webpack-bundle.js") }
       end
@@ -106,7 +65,7 @@ module ReactOnRails
           allow(Webpacker).to receive_message_chain("manifest.lookup")
             .with("webpack-bundle.js")
             .and_raise(Webpacker::Manifest::MissingEntryError)
-          allow(Utils).to receive(:using_webpacker?).and_return(true)
+          allow(ReactOnRails::WebpackerUtils).to receive(:using_webpacker?).and_return(true)
         end
 
         it { expect(subject).to eq("public/webpack/development/webpack-bundle.js") }
@@ -125,7 +84,7 @@ module ReactOnRails
           allow(Webpacker).to receive_message_chain("manifest.lookup!")
             .with("webpack-bundle.js")
             .and_raise(Webpacker::Manifest::MissingEntryError)
-          allow(Utils).to receive(:using_webpacker?).and_return(true)
+          allow(ReactOnRails::WebpackerUtils).to receive(:using_webpacker?).and_return(true)
         end
 
         it { expect(subject).to eq("public/webpack/development/webpack-bundle.js") }
